Add files via upload
Noticed after I finished this that the ADDRESS table isn't on the ER Diagram. Was this intended to hold all addresses? If so, all address columns will need to be modified. Other than that, just try to look over my definitions and see if you agree. The EQUIPMENT table in particular left me with a lot of uncertainty.